refactor(replication): move delete worker contract into crate
