+++ crates/replication/Cargo.toml
@@ -30,6 +30,7 @@ byteorder.workspace = true
 rmp.workspace = true
 rmp-serde.workspace = true
 rustfs-filemeta.workspace = true
+rustfs-storage-api.workspace = true
 rustfs-utils = { workspace = true, features = ["http"] }
 s3s.workspace = true
 serde.workspace = true
