+++ crates/replication/src/lib.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 pub mod config;
+pub mod delete;
 pub mod mrf;
 pub mod operation;
 pub mod resync;
@@ -20,6 +21,7 @@ pub mod rule;
 pub mod tagging;
 
 pub use config::{ObjectOpts, ReplicationConfigurationExt};
+pub use delete::DeletedObjectReplicationInfo;
 pub use mrf::{MrfOpKind, MrfReplicateEntry, decode_mrf_file, encode_mrf_file};
 pub use operation::{MustReplicateOptions, is_ssec_encrypted};
 pub use resync::{
